Set use_cache=False for the model because the cache isn't needed for generating embeddings.
Started setting the fingerprint in `map` because otherwise it appears to try to hash `compute_embeddings` which contains the entire model!

Started moving the output tensors to the CPU in compute_embeddings to avoid running out of VRAM.

Ran experiments to conclude that batched computation is (much!) slower for big models too.